+++ src/app/api/fitbit/import/csv/route.ts
@@ -2,7 +2,7 @@ import { Hono } from "hono";
 import { handle } from "hono/vercel";
 import { FitbitAPI } from "@/lib/fitbit";
 
-const app = new Hono().basePath("/api/fitbit/import");
+const app = new Hono().basePath("/api/fitbit/import/csv");
 
 interface ImportResponse {
   success: boolean;
@@ -14,7 +14,7 @@ interface ImportResponse {
   };
 }
 
-app.post("/csv", async (c) => {
+app.post("/", async (c) => {
   try {
     const body = await c.req.json();
     const { csvContent } = body;
@@ -59,44 +59,6 @@ app.post("/csv", async (c) => {
       success: false,
       message: `Failed to import CSV data: ${errorMessage}`
     } as ImportResponse, 500);
-  }
-});
-
-app.post("/validate", async (c) => {
-  try {
-    const body = await c.req.json();
-    const { csvContent } = body;
-
-    if (!csvContent || typeof csvContent !== 'string') {
-      return c.json({
-        success: false,
-        message: "CSV content is required"
-      } as ImportResponse, 400);
-    }
-
-    const fitbit = new FitbitAPI();
-    
-    // Parse and validate CSV data
-    const csvData = fitbit.parseCSVData(csvContent);
-    
-    return c.json({
-      success: true,
-      message: `CSV validation successful. Found ${csvData.length} valid records`,
-      results: {
-        success: csvData.length,
-        failed: 0,
-        errors: []
-      }
-    } as ImportResponse);
-  } catch (error) {
-    console.error("Error validating CSV data:", error);
-    
-    const errorMessage = error instanceof Error ? error.message : 'Unknown error occurred';
-    
-    return c.json({
-      success: false,
-      message: `CSV validation failed: ${errorMessage}`
-    } as ImportResponse, 400);
   }
 });
